pyproject.toml
[tool.poetry]
name = "fullctl"
repository = "https://github.com/fullctl/fullctl"
version = "0.1.0"
description = "full control"
authors = ["20C <code@20c.com>"]
license = "Apache-2.0"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
]


[tool.poetry.dependencies]
python = "^3.7"
django-handleref = ">=0.5"
djangorestframework = ">=3.11,<4"
Django = ">=2.2,<3"
django-peeringdb = "*"
django-reversion = "<4"
django-inet = "*"
django-autocomplete-light = ">=3,<=4"
social-auth-app-django = "<4"
grainy = ">=1.6.0,<2"
django-grainy = ">=1.9.0,<2"
pyyaml = "*"
pip = "*"
celery = "<6,>=5"
peeringdb = { url = "https://github.com/peeringdb/peeringdb-py/archive/july_updates.zip"}

[tool.poetry.dev-dependencies]
# tests
pytest = ">=6.0.1"
pytest-django = ">=3.8.0"
pytest-cov = "*"
pytest-pythonpath = "*"
# linting
bandit = "^1.6.2"
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
